requirements.txt
langchain
langchain-core
langchain-community
pypdf
sentence-transformers
faiss-cpu
langchain-google-genai
python-dotenv
langchain-text-splitters
langchain-huggingface